+++ src/index.d.ts
@@ -1,0 +1,2 @@
+declare const plugin: {}
+export default plugin